add superposition_circuit and create controller folder
